Finish styling and add ability to change category
