+++ src/API/api.js
@@ -1,4 +1,13 @@
-import API from '../config/config';
+import {
+  loadingMeal, errorMeal, successMeal, loadingMeals, errorMeals, successMeals,
+} from '../redux/actions/index';
+
+const API = {
+  BASE_URL: 'https://www.themealdb.com/api/json/v1/1',
+  FILTER_ENDPOINT: 'filter.php',
+  LOOKUP_ENDPOINT: 'lookup.php',
+  SEARCH_ENDPOINT: 'search.php',
+};
 
 const fetchMeals = async (category) => {
   const url = `${API.BASE_URL}/${API.FILTER_ENDPOINT}/?c=${category}`;
@@ -36,4 +45,40 @@ const fetchMeal = async (id) => {
   throw Error(404);
 };
 
-export { fetchMeals, fetchSearchMeals, fetchMeal };
+const fetchMealsByCategory = (category) => async (dispatch) => {
+  dispatch(loadingMeals());
+  const response = fetchMeals(category);
+  try {
+    const data = await response;
+    return dispatch(successMeals(data.meals));
+  } catch (e) {
+    return dispatch(errorMeals('Error while fetching data.'));
+  }
+};
+
+const searchMeals = (query) => async (dispatch) => {
+  dispatch(loadingMeals());
+  const response = fetchSearchMeals(query);
+  try {
+    const data = await response;
+    if (data.meals) {
+      return dispatch(successMeals(data.meals));
+    }
+    return dispatch(errorMeals(`There is no meal with name: ${query}`));
+  } catch (e) {
+    return dispatch(errorMeals('Error while fetching data.'));
+  }
+};
+
+const fetchMealById = (id) => async (dispatch) => {
+  dispatch(loadingMeal());
+  const response = fetchMeal(id);
+  try {
+    const data = await response;
+    return dispatch(successMeal(data.meals[0]));
+  } catch (e) {
+    return dispatch(errorMeal('Error while fetching data.'));
+  }
+};
+
+export { fetchMealsByCategory, searchMeals, fetchMealById };
