+++ src/index.css
@@ -5,6 +5,7 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
+  background-color: #121a1b;
 }
 
 code {
@@ -61,12 +62,14 @@ code {
 
 .ingredients {
   margin-top: 10px;
+  color: #fff;
 }
 
 .heading {
   border-bottom: 1px solid #ccc;
   padding-bottom: 10px;
   margin-bottom: 10px;
+  color: #fff;
 }
 
 
@@ -74,6 +77,7 @@ code {
 .instructions {
   margin-top: 2rem;
   color: #445052;
+  color: #fff;
 }
 
 .items {
