--- src/redux/reducers/__tests__/categories.test.js
@@ -1,7 +1,0 @@
-import categories from '../categories';
-
-describe('categories reducer', () => {
-  test('should return initial state', () => {
-    expect(categories(undefined, {})).toEqual([]);
-  });
-});
